Rename transformations -> transform, make a dedicated folder
